refactor: move migrate_databases to core helpers
commands.py should be used for cli commands
